[style] integration / socketを使うテスト(1,2,3,4),socketが不要なテスト(5,8),content-typeを見るテスト(6,7)の3つにファイル分け
